+++ src/route/api/account/mainCharacter_PUT.ts
@@ -1,7 +1,7 @@
 import { jsonEndpoint } from '../../../infra/express/protectedEndpoint';
 import { dao } from '../../../db/dao';
-import { idParam } from '../../../route-helper/paramVerifier';
-import { verify, number } from '../../../route-helper/schemaVerifier';
+import { idParam } from '../../../util/express/paramVerifier';
+import { verify, number } from '../../../util/express/schemaVerifier';
 import { canDesignateMain } from '../../../domain/account/canDesignateMain';
 import { CORP_DOOMHEIM } from '../../../shared/eveConstants';
