+++ src/route/api/account/transfer_POST.ts
@@ -1,7 +1,7 @@
 import { jsonEndpoint } from '../../../infra/express/protectedEndpoint';
 import { dao } from '../../../db/dao';
-import { idParam } from '../../../route-helper/paramVerifier';
-import { verify, number } from '../../../route-helper/schemaVerifier';
+import { idParam } from '../../../util/express/paramVerifier';
+import { verify, number } from '../../../util/express/schemaVerifier';
 
 import { BadRequestError } from '../../../error/BadRequestError';
 import { UnauthorizedClientError } from '../../../error/UnauthorizedClientError';
