+++ src/route/api/admin/citadel_POST.ts
@@ -1,7 +1,7 @@
 import { jsonEndpoint } from '../../../infra/express/protectedEndpoint';
 import { DEFAULT_NUM, toNum } from '../../../db/tnex';
 import { dao } from '../../../db/dao';
-import { verify, string, boolean, SchemaVerificationError, } from '../../../route-helper/schemaVerifier';
+import { verify, string, boolean, SchemaVerificationError, } from '../../../util/express/schemaVerifier';
 
 
 export class Input {
