+++ src/route/api/killmail/killmail_GET.ts
@@ -1,7 +1,7 @@
 import { jsonEndpoint } from '../../../infra/express/protectedEndpoint';
 import { Tnex } from '../../../db/tnex';
 import { AccountPrivileges } from '../../../infra/express/privileges';
-import { idParam } from '../../../route-helper/paramVerifier';
+import { idParam } from '../../../util/express/paramVerifier';
 import { dao } from '../../../db/dao';
 import { NotFoundError } from '../../../error/NotFoundError';
 import { ZKillmail } from '../../../data-source/zkillboard/ZKillmail';
