+++ src/route/api/srp/loss/loss_PUT.ts
@@ -1,5 +1,5 @@
 import { jsonEndpoint } from '../../../../infra/express/protectedEndpoint';
-import { number, verify, string, nullable, stringEnum } from '../../../../route-helper/schemaVerifier';
+import { number, verify, string, nullable, stringEnum } from '../../../../util/express/schemaVerifier';
 import { AccountSummary } from '../../../../infra/express/getAccountPrivs';
 import { AccountPrivileges } from '../../../../infra/express/privileges';
 import { Tnex } from '../../../../db/tnex';
@@ -7,7 +7,7 @@ import { dao } from '../../../../db/dao';
 import { BadRequestError } from '../../../../error/BadRequestError';
 import { SrpVerdictStatus, SrpVerdictReason } from '../../../../db/dao/enums';
 import { NotFoundError } from '../../../../error/NotFoundError';
-import { idParam } from '../../../../route-helper/paramVerifier';
+import { idParam } from '../../../../util/express/paramVerifier';
 
 export class Input {
   verdict = stringEnum<SrpVerdictStatus>(SrpVerdictStatus);
