+++ src/route/api/srp/loss/loss_dir_GET.ts
@@ -4,7 +4,7 @@ import { AccountPrivileges } from '../../../../infra/express/privileges';
 import { dao } from '../../../../db/dao';
 import { SrpVerdictStatus } from '../../../../db/dao/enums';
 import { SimpleNumMap, nil } from '../../../../util/simpleTypes';
-import { boolQuery, intQuery, enumQuery } from '../../../../route-helper/paramVerifier';
+import { boolQuery, intQuery, enumQuery } from '../../../../util/express/paramVerifier';
 import { fetchEveNames } from '../../../../data-source/esi/names';
 import { srpLossToJson } from '../../../../domain/srp/srpLossToJson';
 import { SrpLossFilter, SrpLossRow } from '../../../../db/dao/SrpDao';
