+++ src/route/api/srp/loss/triage_GET.ts
@@ -4,7 +4,7 @@ import { AccountPrivileges } from '../../../../infra/express/privileges';
 import { Tnex } from '../../../../db/tnex';
 import { dao } from '../../../../db/dao';
 import { NotFoundError } from '../../../../error/NotFoundError';
-import { idParam } from '../../../../route-helper/paramVerifier';
+import { idParam } from '../../../../util/express/paramVerifier';
 import { SrpTriageJson } from '../../../../domain/srp/SrpLossJson';
 import { triagedLossesToSuggestionJson } from '../../../../domain/srp/triage/triagedLossesToSuggestionJson';
 import { triageLosses } from '../../../../domain/srp/triage/triageLosses';
