+++ src/route/api/srp/payment/payment_GET.ts
@@ -3,7 +3,7 @@ import moment = require('moment');
 import { jsonEndpoint } from '../../../../infra/express/protectedEndpoint';
 import { Tnex } from '../../../../db/tnex';
 import { AccountPrivileges } from '../../../../infra/express/privileges';
-import { idParam } from '../../../../route-helper/paramVerifier';
+import { idParam } from '../../../../util/express/paramVerifier';
 import { dao } from '../../../../db/dao';
 import { NotFoundError } from '../../../../error/NotFoundError';
 import { SimpleNumMap } from '../../../../util/simpleTypes';
